requirements.txt
# install scipy
numpy
scipy
matplotlib

# latest ipython 6.0+ does not support python 2.7
ipython==5.3.0

jupyter
pandas
sympy
nose

# install pillow for preprocessing other than png image
pillow

opencv-python

# from docker image
# docker pull udacity/carnd-term1-starter-kit

# opencv3, same as opencv-python
scikit-learn
scikit-image

h5py
eventlet
flask-socketio
seaborn

ffmpeg-normalize
imageio==2.1.2
pyqt5
moviepy
tensorflow-gpu
tensorflow==0.12.1
keras==1.2.1

# extra
requests


# faster_rcnn
# https://stackoverflow.com/questions/45018492/include-whl-installation-in-requirements-txt
http://download.pytorch.org/whl/cu80/torch-0.3.1-cp27-cp27mu-linux_x86_64.whl ; sys_platform == "linux2"
http://download.pytorch.org/whl/cu80/torch-0.3.1-cp36-cp36m-linux_x86_64.whl ; sys_platform == "linux"
http://download.pytorch.org/whl/torch-0.3.1-cp36-cp36m-macosx_10_7_x86_64.whl ; sys_platform == "darwin"
cupy ; sys_platform == "linux2"
cupy ; sys_platform == "linux"
chainer ; sys_platform == "linux2"
chainer ; sys_platform == "linux"
visdom
torchvision
scikit-image
tqdm
fire
pprint
matplotlib
ipdb
cython
git+https://github.com/pytorch/tnt.git@master


# visualize the graph for pytorch
# graphviz needs install using `brew install graphviz`
# https://github.com/szagoruyko/pytorchviz
git+https://github.com/szagoruyko/pytorchviz

# tensorboard-pytorch
# git+https://github.com/lanpa/tensorboard-pytorch
